feat: expose path param to iac job
The job here is intended as a simple way of interfacing with the IaC
test command, while the scan command can be used for the kitchen sink
approach. This probably needs refining a bit in the future.

This PR exposes the path param at the job level, as without it the
command is a little _too_ minimal. The rest of the config is left at the
command level for granular control