chore: Split scan command into reusable parts
Scan command does a lot apart from running scan:

 1. it installs snyk CLI tool and caches current version,
 2. it configures snyk, and finally
 3. it runs one snyk command.

And while this covers the simplest scenarios, it falls short when we
would like to call snyk more than once.

Changes in this pull request move the installation and configuration
steps into separate commands that developers can use on their own.

Do note that we left the scan command intact: it behaves exactly as it
behaved before our changes (it still performs all steps listed above).
We did this because we do not want to break existing CI pipelines.